Move application generator naming and database code to concerns
